backend/pyproject.toml
[tool.poetry]
name = "tax-lien-strategist-backend"
version = "0.1.0"
description = "FastAPI backend for the Tax Lien Strategist platform"
authors = ["Pryceless Ventures LLC"]
readme = "../README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
alembic = "^1.13.1"
pgvector = "^0.2.4"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.1"
redis = "^5.0.3"
celery = {version = "^5.3.6", extras = ["redis"]}
structlog = "^24.1.0"
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
openai = "^1.14.3"
transformers = "^4.39.3"
tenacity = "^8.2.3"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
httpx = {extras = ["http2"], version = "^0.27.0"}
ruff = "^0.4.2"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
